Add a ReferrerPolicy field to blink.mojom.DragData
A network.mojom.ReferrerPolicy field is added to DragData and
blink::WebDragData, in the same way content::DropData has one. This
prepares for a future in which DropData is replaced by DragData and
allows the migration of DragHostMsg_StartDragging without adding an
extra parameter for the policy.

Bug: 1039255
Change-Id: Ief3493d8fad384dec8df58cb9c38602e29cd4413
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339742
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Commit-Position: refs/heads/master@{#797346}