Store delegated ink points and prepare to draw trail with Skia
Flesh out the delegated ink renderer class a bit more, including a
specific Skia flavor of it, with the expectation that others (software,
GL) may come in the future. Storing the delegated ink points in this
class as they arrive from the browser process, forwarding the metadata
to this class from the aggregated frame, and filtering the points in
the function that will eventually contain the drawing code.

Bug: 1052145
Change-Id: Ic86e37d9ed1d24b6427fa6896b668de8d0402b8e
Cq-Do-Not-Cancel-Tryjobs: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2296567
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Reviewed-by: weiliangc <weiliangc@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#797347}