[FS] Refactor how directory read access is granted.
This gets rid of the redundant ConfirmDirectoryReadAccess method,
and instead just requests read access on the corresponding permission
grant instead. In the future this will make it easier to support a
single prompt for getting access to a writable directory, instead
of the current two separate prompts.

To do this, also adds an extra flag to the
NFSPermissionGrant::RequestPermission method to indicate if this
particular permission requests requires user activation.

Bug: 1115632
Change-Id: Ie3980867d162e16723569a0fd36f27323d2700ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2341214
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797406}