Disable hand pointer on hovering disabled toggle
The 'actionable' attribute was added to 'property-box' in
network_config_toggle.html. This caused a pointer cursor to
be displayed when the element was disabled.

This change removes the 'actionable' property and adds css to change
the cursor to initial if the element is disabled and pointer
if it is enabled.

Bug: 1114668
Change-Id: If09818400ae37b3f3c21f15b161d37940ae4ed62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2347196
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Claude van der Merwe <cvandermerwe@google.com>
Cr-Commit-Position: refs/heads/master@{#797425}