Call UpdateTexImage after OnFrameAvailable
Both with SurfaceTexture and AImageReader we get only one FrameAvailable
callback before we call UpdateTexImage. This causes the problem when we
released buffer to codec in early rendering, but dropped the frame
later. In this case we did not called UpdateTexImage for that frame and
will not received OnFrameAvailable for next one causing
WaitForFrameAvailable to timeout.

This CL adds UpdateTexImage call after WaitForFrameAvailable in
RenderToTextureOwnerBackBuffer in case we're already waiting for frame.

Bug: 1113203
Change-Id: I32e703d07df94f7170795077773a30b4578c1a86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339495
Commit-Queue: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797477}