[omnibox] Add kFocusGestureTriggersContextualWebZeroSuggest flag
This CL adds an enabled-by-default flag that causes on-focus gestures
to trigger ZeroSuggest on the Contextual Web page classification
(OTHER).

We add this flag so we can disable it in some experiments, where we
want to use on-clobber instead.

It also adds a test to make sure this flag doesn't affect other page
classifications such as the NTP.

Bug: 1106096
Change-Id: I32d4b9fc396874f836c60ae347ac3894818cf1d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350309
Commit-Queue: Tommy Li <tommycli@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797491}