Split token cache from IdentityAPI and add tests
Implementing a subset matching cache will require a more complex cache
for the IdentityAPI class. At the moment, the existing cache is
contained inside the IdentityAPI class.

This CL separates the two, so all of the caching logic for IdentityAPI
is in its own class. Additionally, unit tests have been added to test
the basic functionality of this cache.

Bug: 1100535
Change-Id: I39f5357d3ce87ae4d958aa2c8585cc2deec624be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2343318
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: William Lin <williamlin@google.com>
Cr-Commit-Position: refs/heads/master@{#797527}