Loading Predictor: Disallow prefetches to private IP addresses.
This CL blocks prefetching to private IP addresses, so that CORS-RFC1918
is not bypassed. It adds a URLLoaderFactory option to accomplish this.
We can't just use CORS-RFC1918 as-is because the security state of the
client and in fact the client's URL is not known at prefetching time,
as redirects can happen after this time.

Bug: 1114064
Change-Id: I74bdbb078d3b3ed41e9fbbc809f68b43499b6040
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351102
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797559}