Don't rethrow exceptions in the chrome://media-app MessagePipe.
Rethrowing MessagePipe exceptions only makes sense when a particular
side of the pipe should _never_ throw an exception. Exceptions that
propagate to the other side to be handled should not be re-thrown on the
server side.

Currently any trusted-context exception will trigger a crash report due
to new handling for the `unhandledrejection` event that was added in
r790636.

Bug: b/163927591, b/158135844
Change-Id: I0d01058b2e34dbf2742ddd05ce54a45007863f52
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352582
Reviewed-by: Patti <patricialor@chromium.org>
Commit-Queue: Trent Apted <tapted@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797585}