Kill renderer if message arrives while page is in bfcache
Renderers should not be sending any messages through an associated
interface if the frame is in the bfcache. We can assume that if such a
thing happens the renderer might have been compromised, so kill it.

We can only do this reliably for associated interfaces as we can
guarantee that after the SetPageLifecycleState ack arrives no more
messages should due to the ordering guarantees.

Gated with a feature parameter so that we can roll out slowly (and
potentially catch renderer bugs).

Bug: 1111720

Change-Id: Ib8f0cc967639a2808f4e195f91fce34a58c1d203
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2312821
Commit-Queue: Carlos Caballero <carlscab@google.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797599}