[PwdCheckAndroid] Start App over CCT if available
If the password for an App is compromised, the password check now offers
a change button to launch the app (provided the app package is locally
available).
If the app is not installed, a hint informs the user that they can
change the password in the app.

For regular websites, a CCT will navigate to the index page (or
.well-known/change-password if that feature is active).

Screenshot in the bug.

Bug: 1115117, 1092444
Change-Id: I6debb91070de44af62ecd74ca97fb2e4744d6385
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2351782
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797603}