Re-enable UkmBrowserTest.SingleSyncSignOut.
Calling TearDownAuthForTesting() causes a crash in
SigninManager::signOut() (Java). This is because the sign-out operation
started during the test body does not finish before
TearDownAuthForTesting() is called. In other words, we try to sign out
during tear down while sign-out is already in progress.

We don't need to sign out during tear down, so we'll stop doing this.
However, we'll continue to reset the AccountManagerFacadeProvider.

Bug: 1102746
Change-Id: I0cc97c8cfa04591d150305418858eaa0abe4b38e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2350231
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Alice Wang <aliceywang@chromium.org>
Commit-Queue: Caitlin Fischer <caitlinfischer@google.com>
Cr-Commit-Position: refs/heads/master@{#797647}