Make sure tab hover cards aren't shown for a closed tab.
There was a bug that hover cards would flash in when a tab was closed,
this could be due to missing a mouse event that would have updated the
hover card while the delay timer was running. This change adds a check
to make sure the hover card won't be shown if the tab was destroyed
while the timer was running.

Bug: 982964
Change-Id: I02b95f83f28b45138e61c75fe42cd62208248093
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2349922
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Caroline Rising <corising@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797689}