Get the scale from the raw Arc png to generate the matched mask.
For Arc icon, the current implementation is:
1. Call the function "CreateSuperimposedImage" to generate the 'mix'
image with the background and foregorund.
2. Generate the mask image based on the above 'mix' image.
3. Apply the mask image on the 'mix' image to generate the standard
icon.

However, the 'mix' image generated by the operation function
"CreateSuperimposedImage" doesn't have the same ImageSkiaRep, and scale
as the raw image: foreground image, background image. So the generated
mask image has 1 ImageSkiaRep for the scale 1.0f only. Therefore in
ChomeOS, the icon shown uses the ImageSkiaRep with scale 1.0, which is
not a good quality.

This CL refactors the LoadMaskImage to get the scale and size from the
Arc foreground image, not the 'mix' generated image, to create the
matched mask image for Arc icons, so that in the ChromeOS the icon shown
can use the ImageSkiaRep with scale 2.0f which is better quality.

BUG=1114771

Change-Id: I7564d34b0be6950516e3c25480226ecf4aac403a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353972
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797691}