[Live Caption] Inform the UI of an error when the speech service disconnects.
Add disconnect handlers to the speech service. If the UI in the browser
disonnects from the service in the renderer, have the speech service stop
transcribing captions. If the service in the renderer disconnects from
the utility process which connects to SODA or the cloud, send a message
to the UI that there was an error and have the UI display the error
message.

Bug: 1055150
Change-Id: I3940cae2cd12503353e105fd612b064908d9951c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353024
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Evan Liu <evliu@google.com>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/master@{#797773}