chrome://devices: Add link to string
Add the appropriate help link to chrome://devices

Also fixing a typo in the link itself - should be support.google.com,
not www.google.com.

Note: String itself is not modified by this change, but is detected
as a change due to adding the URL argument.

Bug: 1114186
Change-Id: Ibb8583d419639da3b1ba13058060cf91298deeb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353406
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797785}