Only high res tiles can use LCD text
Tiles with different scale can't use LCD text because their pixels
are not aligned with physical pixels. Once the flag is set on a tiling,
it should keep unchanged and set the flag for all tiles of the tiling,
to avoid mosaic effect caused by different LCD text settings of
different tiles, and we accept the temporary imperfection of LCD-text.

Also when a layer's LCD-text status changes, only invalidate the highres
tiling in the same way as we invalidate the highres tiling when raster
translation changes.

This is a preparation CL for crbug.com/1111195.

Bug: 1111195
Change-Id: I64ef167ac322ee94c96a82575e3f2448060fcda6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2344809
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: vmpstr <vmpstr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797793}