[M85] RFHI: Refresh feature policy after learning a new frame's origin
This is a merge of a fix to a long-standing bug with inconsistent
Feature Policy state between the renderer and the browser in some same-
origin subframes that have not committed navigations. It manifested
itself during a gradual rollout of a feature in M84. We mitigated the
resulting crashes in M84 with a change partly disabling a server-side
rollout config, and got approval to merge the fix to M85 in order to
unblock the rollout.

Merge approval: https://crbug.com/1111735#c43
Deltas with original CL: only tests, to accommodate some M85->M86
differences in test scaffolding and Trust Token API syntax. Confirmed
locally that the test fails without the functional change in place.

-- Original CL message:

The browser's views of certain same-origin subframes' feature policies
can disagree from the values available to a well-behaved renderer,
erroneously triggering this renderer kill. This is because not all
same-origin subframes commit navigations, but the Feature Policy code in
RenderFrameHostImpl depends on a call to DidNavigate to update RFHI's
|feature_policy_| member after learning a new frame's origin. For frames
that don't see DidNavigate calls, RFHI::feature_policy_ ends up in a
state that can disagree with the renderer-side feature policy state for
the same frame, stored in blink::SecurityContext.

This change adds a fix by adding a call to ResetFeaturePolicy to
RFHI::SetOriginAndIsolationInfoOfNewFrame, which is the first time that
the browser's representation of a new frame learns the frame's origin.
This allows computing a correct feature policy for features whose
values depend on the frame origin. (Currently, we compute the policy
before knowing the origin, which leads to the problematic inconsistent
state.)

(cherry picked from commit 35460b0e4f4262341af56d699f9e4512fc0d8f74)

Bug: 1112959, 1111735
Test: Add a browser test that crashes before the fix and passes after
Change-Id: Ic3c42fbf10a468d230d85981eba10b4be4533307
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335910
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Charlie Hu <chenleihu@google.com>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#795625}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354514
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#1481}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}