Define XrDestroySession for FakeOpenXrImpl
Some tests were currently failing on code that (seemed to be) trying to
call XrDestroySession, which was not defined by the FakeOpenXrImpl API.
This change defines that method, so the tests that were disabled should
be passing again. If the flakes still occur, some logging will need to
be upgraded to DLOG(ERROR), so that we can get a better idea of what's
going on, since there is some amount of distrust of the callstacks.

Fixed: 1115721
Change-Id: Ibeb8fdeb38c84da83be886cf6a02b53289d76b89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354523
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797907}