Switch Access: Ensure user can cancel when deactivating switch access.
This change ensures that switch access is in a stable state if the user
cancels when disabling it. Previously, pressing cancel
triggered a buggy re-initialization of switch access that resulted in
disabed switches. This change simply turns switch access off before
re-activating it to ensure it's in a stable state. Though we do turn
switch access off, then on again, the end result for the user is that
switch access stays on.

Bug: 1115126
AX-Relnotes: N/A

Change-Id: I57c37e02122f137f64c38e57aff3009678aec2b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354622
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Cr-Commit-Position: refs/heads/master@{#797927}