Update It2MeHost to use RemoteAccessHostAllowRelayedConnection policy
This CL adds some logic to the It2MeHost to read this policy and update
the network settings to reflect its value.  The network settings are
used to discard the entries in the ICE Config which are blocked by
policy.  With this change, the Me2MeHost and It2MeHost will now have
the same behavior (previously the It2MeHost only read the NatTraversal
policy).

This should be a safe change as the AllowRelayedConnection policy
defaults to true if not explicitly set and the description of the policy
does not indicate that it only applies to Me2Me.  Another interesting
note is that we do not refer to it in our Enterprise config doc (only
the firewall traversal policy is mentioned) so I would be surprised
if it is broadly used.

Once this lands, I'll update the website to use the new
'relayConnectionsAllowed' message value.

Change-Id: Ide4009e5e7979f7bda6dc098e5968739ed50e97e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353931
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797935}