ash: Promote AuthDialogDebugView to be default view
AuthDialogDebugView was used in in-session auth dialog in the early
development stage. Now that it has the correct views and is close to
the final contents view, rename it to AuthDialogContentsView. This
change also removes the dev flag "show-auth-dialog-dev-overlay" so that:

1. developers (especially Chrome's WebAuthn team) can test the flow end to
   end without having to enter dev mode on ChromeOS.
2. UX reviewers can evaluate the user flow.

Note that the feature is still behind a feature flag "base::Feature
kWebAuthCrosPlatformAuthenticator". Developers can toggled that via
chrome://flags.

No functionality change.

Bug: b:156258540
Change-Id: I7c8e0a94993cc6d483d5280cd91fa970846930f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2355000
Commit-Queue: Yicheng Li <yichengli@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#797959}