serial: Ensure close callback is always executed
This change fixes SerialIoHandler::Close() to ensure that the callback
is always executed even if there isn't a file handle to close. Failing
to do so will trigger a DCHECK in the Mojo bindings layer and could
cause clients of the Mojo interface to stall waiting for a response.

Bug: 884928
Change-Id: I272cc1987627232dc2143380071a07b7c58fb2cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354775
Auto-Submit: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798115}