[BackgroundFetch] Fix image resource type regression.
ImageResource's type members is a mojo_base.mojom.String16? which is
typemapped to WTF::String in blink. For some reason though, passing a
null WTF::String to mojom is crashing the renderer.

This CL explicitly sets the image resource type to the empty string if
not available, and adds a web test to make sure this doesn't happen
again.

Change-Id: I115666b6e42011dfc398965bd0f4e795d1869979
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352878
Commit-Queue: Rayan Kanso <rayankans@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798134}