Fade shelf out first in UnlockAnimationBeforeUIDestroyed
When running UnlockAnimationBeforUIDestroyed, animate shelf before
lock screen containers. Shelf fades out and then fades in when the
unlock animation completes - i.e. in the
UnlockAnimationBeforeUIDestroyed callback. If animations run
immediately, the fade out and fade in animation order could get reversed
if fade out animation is started after the part of unlock animation to
which animation callback is passed.

I'm not sure that this is actually the root cause for the linked issue,
but it is a way to get to an invisible shelf (it requires zero length
animation, which may happen during screen rotation).

BUG=1106796

Change-Id: I8f26200666e742b249ba96b64a32ce28022634f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352905
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Toni Baržić <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798145}