Updating 'supported on' value for RemoteAccessHostAllowRelayConnections
The current value for this policy indicates that it applies to all
ChromeOS versions after M41 however this isn't right as that policy was
not read on ChromeOS prior to a change I made in M86.  This CL updates
the 'supported on' value to match what the product actually supports.

BYPASS_POLICY_COMPATIBILITY_CHECK=The previous value was incorrect but
setting the correct value causes the policy checker to complain.

Change-Id: I50e21fe9828cb54f6cfaa11a9a120c0f4b89ad40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354593
Commit-Queue: Joe Downing <joedow@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798166}