Allow disabling StartupTabPreloader from the Intent.
The AGA experiment id changes in:
  https://chromium-review.googlesource.com/c/chromium/src/+/2338934
does not work when the StartupTabPreloader is used. Since the
StartupTabPreloader launches the navigation before all of native is
loaded, fixing this isn't simply a matter of moving the experiment
reading code earlier - it would involve changing Chrome's startup logic.

This change allows clients to disable the StartupTabPreloader to ensure
that the experiment logic in the linked CL is triggered (at the cost of
a longer time to navigation on cold start).

Change-Id: Ie83b4f38705084c567d7c20821600de7cdf15695
Bug: 1111941
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2356505
Commit-Queue: Peter Conn <peconn@chromium.org>
Commit-Queue: Yaron Friedman <yfriedman@chromium.org>
Auto-Submit: Peter Conn <peconn@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798203}