IsolatedWorldCSP: Plumb DOMWrapperWorld for all ResourceLoaderOptions.
This is part of the work to enforce isolated world CSP for resource
requests when the feature IsolatedWorldCSP is enabled and to correctly
bypass the main world CSP otherwise.

Note that it's possible that the plumbed DOMWrapperWorld might not
always be completely correct (for example if it is plumbed
asynchronously from the point which caused the resource request).
However this is still better than the status quo where the CSP checks
for isolated worlds are bypassed based on calling
FrameFetchContext::ShouldBypassMainWorldCSP() which is often incorrect
especially for redirects since we might be in a different different
world by then.

Next CL will remove FetchContext::ShouldBypassMainWorldCSP.

BUG=1099975

Change-Id: I727513d0fa8e9e16aa05a0fdd7c86d987f983890
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354566
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798215}