Stop excluding clips when computing filter bounding boxes
The code was added not for this use case, but to ensure that
composited bounds were optimized to avoid parts of descendants
that were clipped out.

This change also unblocks deleting some code and removing one
cache slot, which will save memory on PaintLayer.

Change-Id: I0ca7592ab3609a784bedd095e0c7fb4315b7527a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353872
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798251}