Modify observe all window to observe browser tabs only.
Sometimes, when shutdown the system, the instance could be updated, but
the window might have been destroyed. Also ARC CTS might have the
unusual event flow to set the task active when the window is killed. So
observing all windows when update the instance could cause crash.

Actually CL:2307669 added observing all windows mainly because browser
tabs are not observed. So redo CL:2307669:
1. Observe browser tabs during the system init phase
2. Observe the new inserted tab

To avoid crash, OnInstance checks whether the window is observed, and
if not, remove the instance.

BUG=1116041
BUG=1116033
b:163005244

Change-Id: Idfb675b61592176da183be092dcecff1e464ec2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2355163
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798263}