macviews: fix NativeWidgetMacTest.NonWidgetParentLastReference on Big Sur
This test asserts that closing a parent NSWindow deallocates the child
NSWindow if the parent held the last reference. This is all well and
good, but needs to account for AppKit's habit of autoreleasing refs
to things while operating on them (eg, https://crbug.com/998773). Do this
by allowing the autorelease pool to drain before checking for dealloc.

Bug: 1114245
Change-Id: I46426f163b5fbdc3822ec4c4019c722cac787c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2356910
Commit-Queue: Robert Liao <robliao@chromium.org>
Reviewed-by: Robert Liao <robliao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798275}