Add site-/origin-keying bit to SiteInfo.
The fourth (and final) CL in implementing origin-/site-keying of
SiteInstances. https://crbug.com/1085275/#c2 lists the sequence of CLs
and what each will accomplish.

This CL adds the keying bit to SiteInfo, and updates MakeTie() and
ComputeSiteInfo() to use it. It removes the prior hack in GetSiteForURL
regarding isolated base origins, and adds/updates tests that involve
isolated base origins.

This CL also changes the order in which we check isolated origins, so
that opt-in isolation requests take precedence over command-line
isolation.

Bug: 1067389, 1085275
Change-Id: I3f91fe94a1a61a7ef272d83232a6ff4e6e3bff3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269516
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#798294}